configure entry point for parcel
