+++ app/index.html
@@ -1,11 +1,12 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-     <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.2.11/semantic.min.css"></link> 
+     <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.2.11/semantic.min.css"></link>
     <meta charset="UTF-8">
     <title>Solid Quotes</title>
   </head>
   <body>
     <div id="app"></div>
+    <script src="./index.js"></script>
   </body>
 </html>
