+++ app/index.js
@@ -1,0 +1,12 @@
+//package dependicies
+import 'babel-polyfill';
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+//local dependicies
+import App from './app.jsx'
+
+ReactDOM.render(
+  <App />,
+  document.getElementById('app')
+);
